+++ cansocket/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix frame send data length
 - Use libc constants for error frame checking
 - Fix `Socket` doc test
+- Added option to enable protocol errors
 
 ## v0.1.0